+++ simple-chat/src/components/chat.js
@@ -1,0 +1,5 @@
+export function goBackToChatList(chatListScreen, chatScreen, setCurrentChatId) {
+    chatScreen.style.display = 'none';
+    chatListScreen.style.display = 'block';
+    setCurrentChatId(null);
+}
