+++ simple-chat/src/components/switch.js
@@ -1,0 +1,9 @@
+export function switchUser(currentUser, currentChatName, isSwitched) {
+    if (!isSwitched) {
+        currentUser = currentChatName;
+    } else {
+        currentUser = 'Максим';
+    }
+    isSwitched = !isSwitched;
+    return { currentUser, isSwitched };
+}
